CmuxGit stage 2: extract the GitHub PR probe pipeline from TabManager
Lift sub-domain B into PullRequestProbeService, a stateless Sendable struct
(same shape as GitMetadataService: nonisolated async, off the caller's actor,
parallel): candidate-seed slug resolution (via GitMetadataService), concurrent
REST repo fetch + per-branch lookups with the caller-owned repo cache,
GH_TOKEN/`gh auth token` auth via injected CommandRunning, and the pure
selection/policy logic (preferred PR, stale-merged filtering, cache/refresh
policy). Value types (WorkspacePullRequest* family, GitHubPullRequestProbeItem)
move to the package; new PullRequestStatus enum bridges the app's
SidebarPullRequestStatus by raw value, so the apply path is unchanged.

TabManager keeps the orchestration (probe state machines, per-key dictionaries,
poll timers, the apply path, and ownership of the repo cache) and calls the
service; commandRunner stays an init parameter and now feeds the service; a
debugLog closure seam carries cmuxDebugLog in DEBUG builds. Another -840 lines
(11,506 -> 9,557 across both stages). The DispatchSourceTimer poll loops are
deliberately NOT redesigned in this PR (existing app-target orchestration;
timing changes need their own dogfood).

Tests: the 7 pure TabManagerPullRequestProbeTests migrate to the package
(48 package tests total), plus new REST-decode (merged_at -> MERGED synthesis),
branch-endpoint encoding, and result-resolution coverage. The old "not a url"
malformed-URL fixture only failed URL(string:) on pre-macOS-14 SDK parsing;
replaced with an empty string, malformed on every SDK. ci.yml's package-test
list gains CmuxFileWatch, CmuxGit, and CmuxProcess - their suites were compiled
but never executed in CI (each verified passing standalone).

swift build + 48 swift tests, cmux, and cmux-unit all green.
