+++ Packages/CmuxGit/Sources/CmuxGit/Model/WorkspacePullRequestBranchFetchResult.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// The outcome of one per-branch pull-request lookup.
+enum WorkspacePullRequestBranchFetchResult: Sendable {
+    case found(GitHubPullRequestProbeItem)
+    case notFound
+    case transientFailure
+}
