+++ Packages/CmuxGit/Sources/CmuxGit/Model/WorkspacePullRequestBranchLookupOutcome.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// The cache entry produced by a batch of per-branch lookups, plus the branches
+/// that failed transiently.
+struct WorkspacePullRequestBranchLookupOutcome: Sendable {
+    let cacheEntry: WorkspacePullRequestRepoCacheEntry
+    let transientBranches: Set<String>
+}
