+++ Packages/CmuxGit/Sources/CmuxGit/Model/WorkspacePullRequestHTTPResponse.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// A GitHub API response reduced to status code + body.
+struct WorkspacePullRequestHTTPResponse: Sendable {
+    let statusCode: Int
+    let data: Data
+}
