+++ cmuxTests/TabManagerUnitTests.swift
@@ -614,169 +614,8 @@ final class TabManagerWorkspaceOwnershipTests: XCTestCase {
 @MainActor
 final class TabManagerPullRequestProbeTests: XCTestCase {
 
-    func testPreferredPullRequestPrefersOpenOverMergedAndClosed() {
-        let candidates = [
-            TabManager.GitHubPullRequestProbeItem(
-                number: 1889,
-                state: "MERGED",
-                url: "https://github.com/manaflow-ai/cmux/pull/1889",
-                updatedAt: "2026-03-20T18:00:00Z"
-            ),
-            TabManager.GitHubPullRequestProbeItem(
-                number: 1891,
-                state: "OPEN",
-                url: "https://github.com/manaflow-ai/cmux/pull/1891",
-                updatedAt: "2026-03-19T18:00:00Z"
-            ),
-            TabManager.GitHubPullRequestProbeItem(
-                number: 1800,
-                state: "CLOSED",
-                url: "https://github.com/manaflow-ai/cmux/pull/1800",
-                updatedAt: "2026-03-21T18:00:00Z"
-            ),
-        ]
-
-        XCTAssertEqual(
-            TabManager.preferredPullRequest(from: candidates),
-            candidates[1]
-        )
-    }
-
-    func testPreferredPullRequestPrefersMostRecentlyUpdatedWithinSameStatus() {
-        let olderOpen = TabManager.GitHubPullRequestProbeItem(
-            number: 1880,
-            state: "OPEN",
-            url: "https://github.com/manaflow-ai/cmux/pull/1880",
-            updatedAt: "2026-03-18T18:00:00Z"
-        )
-        let newerOpen = TabManager.GitHubPullRequestProbeItem(
-            number: 1890,
-            state: "OPEN",
-            url: "https://github.com/manaflow-ai/cmux/pull/1890",
-            updatedAt: "2026-03-20T18:00:00Z"
-        )
-
-        XCTAssertEqual(
-            TabManager.preferredPullRequest(from: [olderOpen, newerOpen]),
-            newerOpen
-        )
-    }
-
-    func testPreferredPullRequestIgnoresMalformedCandidates() {
-        let valid = TabManager.GitHubPullRequestProbeItem(
-            number: 1888,
-            state: "OPEN",
-            url: "https://github.com/manaflow-ai/cmux/pull/1888",
-            updatedAt: "2026-03-20T18:00:00Z"
-        )
-
-        XCTAssertEqual(
-            TabManager.preferredPullRequest(from: [
-                TabManager.GitHubPullRequestProbeItem(
-                    number: 9999,
-                    state: "WHATEVER",
-                    url: "https://github.com/manaflow-ai/cmux/pull/9999",
-                    updatedAt: "2026-03-21T18:00:00Z"
-                ),
-                TabManager.GitHubPullRequestProbeItem(
-                    number: 10000,
-                    state: "OPEN",
-                    url: "not a url",
-                    updatedAt: "2026-03-21T18:00:00Z"
-                ),
-                valid,
-            ]),
-            valid
-        )
-    }
-
-    func testPullRequestMapDropsStaleMergedHeadPullRequestForLongLivedBaseBranch() throws {
-        let now = try XCTUnwrap(ISO8601DateFormatter().date(from: "2026-04-20T12:00:00Z"))
-        let pullRequests = [
-            TabManager.GitHubPullRequestProbeItem(
-                number: 2400,
-                state: "MERGED",
-                url: "https://github.com/manaflow-ai/cmux/pull/2400",
-                updatedAt: "2026-03-06T12:00:00Z",
-                mergedAt: "2026-03-06T12:00:00Z",
-                headRefName: "develop",
-                baseRefName: "main"
-            ),
-            TabManager.GitHubPullRequestProbeItem(
-                number: 2501,
-                state: "MERGED",
-                url: "https://github.com/manaflow-ai/cmux/pull/2501",
-                updatedAt: "2026-04-19T12:00:00Z",
-                mergedAt: "2026-04-19T12:00:00Z",
-                headRefName: "feature/recent-one",
-                baseRefName: "develop"
-            ),
-            TabManager.GitHubPullRequestProbeItem(
-                number: 2502,
-                state: "OPEN",
-                url: "https://github.com/manaflow-ai/cmux/pull/2502",
-                updatedAt: "2026-04-20T12:00:00Z",
-                headRefName: "feature/recent-two",
-                baseRefName: "develop"
-            ),
-        ]
-
-        let pullRequestsByBranch = TabManager.pullRequestMapByNormalizedBranchForTesting(
-            from: pullRequests,
-            now: now
-        )
-
-        XCTAssertNil(pullRequestsByBranch["develop"])
-        XCTAssertEqual(pullRequestsByBranch["feature/recent-one"]?.number, 2501)
-        XCTAssertEqual(pullRequestsByBranch["feature/recent-two"]?.number, 2502)
-    }
-
-    func testShouldSkipWorkspacePullRequestLookupOnlyForExactMainAndMaster() {
-        XCTAssertTrue(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "main"))
-        XCTAssertTrue(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "master"))
-        XCTAssertTrue(TabManager.shouldSkipWorkspacePullRequestLookup(branch: " master \n"))
-
-        XCTAssertFalse(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "Main"))
-        XCTAssertFalse(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "mainline"))
-        XCTAssertFalse(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "feature/main"))
-        XCTAssertFalse(TabManager.shouldSkipWorkspacePullRequestLookup(branch: "release/master-fix"))
-    }
-
-    func testWorkspacePullRequestRefreshAllowsRepoCacheForTimerAndPeriodicReasons() {
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "periodicPoll"))
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "periodicPoll.followUp"))
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "selectedPeriodicPoll"))
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "selectedPeriodicPoll.followUp"))
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "timer"))
-        XCTAssertTrue(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "timer.followUp"))
-
-        XCTAssertFalse(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "branchChange"))
-        XCTAssertFalse(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "branchChange.followUp"))
-        XCTAssertFalse(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "shellPrompt"))
-        XCTAssertFalse(TabManager.workspacePullRequestRefreshAllowsRepoCache(reason: "commandHint:merge"))
-    }
-
-    func testWorkspacePullRequestShouldRefreshHonorsForcedRefreshForTerminalStates() {
-        let now = Date(timeIntervalSince1970: 1_000)
-        let recentTerminalRefresh = now.addingTimeInterval(-60)
-
-        XCTAssertTrue(
-            TabManager.shouldRefreshWorkspacePullRequest(
-                now: now,
-                nextPollAt: .distantPast,
-                lastTerminalStateRefreshAt: recentTerminalRefresh,
-                currentPullRequestStatus: .merged
-            )
-        )
-        XCTAssertFalse(
-            TabManager.shouldRefreshWorkspacePullRequest(
-                now: now,
-                nextPollAt: now.addingTimeInterval(60),
-                lastTerminalStateRefreshAt: recentTerminalRefresh,
-                currentPullRequestStatus: .closed
-            )
-        )
-    }
+    // Pure pull-request selection/policy tests moved to the CmuxGit package
+    // (CmuxGitTests.PullRequestProbeServiceTests) with the extraction.
 
     func testTrackedWorkspaceGitMetadataPollCandidatesIncludeMainAndMasterPanels() throws {
         let manager = TabManager()
